% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggforest.R
\name{ggforest}
\alias{ggforest}
\title{Forest plot.}
\usage{
ggforest(ma, null_centre = 0, title = NULL)
}
\arguments{
\item{ma}{A meta-analysis object produced by metafor.}

\item{null_centre}{Set the vertical line x-intercept representing null hypothesis for true effect.}

\item{title}{Optional character string for the plot.}
}
\description{
Analogous to the metafor::forest function, but using ggplot.
}
